Move C++ source to a separate file
