+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include LICENSE.md MANIFES.in
+recursive-include *.cpp
